Added int8 version of keras2c using vector processor

+++ apps/keras2c/include/k2c_include.h
@@ -15,16 +15,16 @@ https://github.com/f0uriest/keras2c
 
 
 // Activations
-void k2c_linear_func(float * x, const size_t size);
-void k2c_exponential_func(float * x, const size_t size);
-void k2c_relu_func(float * x, const size_t size);
-void k2c_hard_sigmoid_func(float * x, const size_t size);
-void k2c_tanh_func(float * x, const size_t size);
-void k2c_sigmoid_func(float * x, const size_t size);
-void k2c_softmax_func(float * x, const size_t size);
-void k2c_softplus_func(float * x, const size_t size);
-void k2c_softsign_func(float * x, const size_t size);
-typedef void k2c_activationType(float * x, const size_t size);
+void k2c_linear_func(int8_t  * x, const size_t size);
+void k2c_exponential_func(int8_t  * x, const size_t size);
+void k2c_relu_func(int8_t * x, const size_t size);
+void k2c_hard_sigmoid_func(int8_t  * x, const size_t size);
+void k2c_tanh_func(int8_t  * x, const size_t size);
+void k2c_sigmoid_func(int8_t  * x, const size_t size);
+void k2c_softmax_func(int8_t  * x, const size_t size);
+void k2c_softplus_func(int8_t  * x, const size_t size);
+void k2c_softsign_func(int8_t  * x, const size_t size);
+typedef void k2c_activationType(int8_t  * x, const size_t size);
 extern k2c_activationType * k2c_linear;
 extern k2c_activationType * k2c_exponential;
 extern k2c_activationType * k2c_relu;
@@ -36,23 +36,24 @@ extern k2c_activationType * k2c_softplus;
 extern k2c_activationType * k2c_softsign;
 
 // Advanced Activations
-void k2c_LeakyReLU(float * x, const size_t size, const float alpha);
-void k2c_PReLU(float * x, const size_t size, const float * alpha);
-void k2c_ELU(float * x, const size_t size, const float alpha);
-void k2c_ThresholdedReLU(float * x, const size_t size, const float theta);
-void k2c_ReLU(float * x, const size_t size, const float max_value, const float negative_slope,
-              const float threshold);
+void k2c_LeakyReLU(int8_t  * x, const size_t size, const int8_t  alpha);
+void k2c_PReLU(int8_t  * x, const size_t size, const int8_t  * alpha);
+void k2c_ELU(int8_t  * x, const size_t size, const int8_t  alpha);
+void k2c_ThresholdedReLU(int8_t  * x, const size_t size, const int8_t  theta);
+void k2c_ReLU(int8_t * x, const size_t size, const int8_t max_value, const int8_t  negative_slope,
+              const int8_t  threshold);
 
 // Convolutions
-void k2c_pad1d(k2c_tensor* output, const k2c_tensor* input, const float fill,
+void k2c_pad1d(k2c_tensor* output, const k2c_tensor* input, const int8_t  fill,
                const size_t * pad);
-void k2c_pad2d(k2c_tensor* output, const k2c_tensor* input, const float fill,
+void k2c_pad2d(k2c_tensor* output, const k2c_tensor* input, const int8_t  fill,
                const size_t * pad);
-void k2c_pad3d(k2c_tensor* output, const k2c_tensor* input, const float fill,
+void k2c_pad3d(k2c_tensor* output, const k2c_tensor* input, const int8_t  fill,
                const size_t * pad);
 void k2c_conv1d(k2c_tensor* output, const k2c_tensor* input, const k2c_tensor* kernel,
                 const k2c_tensor* bias, const size_t stride, const size_t dilation,
                 k2c_activationType *activation);
+
 void k2c_conv2d(k2c_tensor* output, const k2c_tensor* input, const k2c_tensor* kernel,
                 const k2c_tensor* bias, const size_t * stride, const size_t * dilation,
                 k2c_activationType *activation);
@@ -68,7 +69,7 @@ void k2c_upsampling3d(k2c_tensor* output, const k2c_tensor* input, const size_t 
 
 // Core Layers
 void k2c_dense(k2c_tensor* output, const k2c_tensor* input, const k2c_tensor* kernel,
-               const k2c_tensor* bias, k2c_activationType *activation, float * fwork);
+               const k2c_tensor* bias, k2c_activationType *activation, int8_t * fwork);
 void k2c_flatten(k2c_tensor *output, const k2c_tensor* input);
 void k2c_reshape(k2c_tensor *output, const k2c_tensor* input, const size_t * newshp,
                  const size_t newndim);
@@ -80,17 +81,17 @@ void k2c_repeat_vector(k2c_tensor* output, const k2c_tensor* input, const size_t
 void k2c_embedding(k2c_tensor* outputs, const k2c_tensor* inputs, const k2c_tensor* kernel);
 
 // Helper functions
-void k2c_matmul(float * C, const float * A, const float * B, const size_t outrows,
+void k2c_matmul(int8_t  * C, const int8_t  * A, const int8_t  * B, const size_t outrows,
                 const size_t outcols, const size_t innerdim);
-void k2c_affine_matmul(float * C, const float * A, const float * B, const float * d,
+void k2c_affine_matmul(int8_t  * C, const int8_t  * A, const int8_t  * B, const int8_t  * d,
                        const size_t outrows,const size_t outcols, const size_t innerdim);
 size_t k2c_sub2idx(const size_t * sub, const size_t * shape, const size_t ndim);
 void k2c_idx2sub(const size_t idx, size_t * sub, const size_t * shape, const size_t ndim);
 void k2c_dot(k2c_tensor* C, const k2c_tensor* A, const k2c_tensor* B, const size_t * axesA,
-             const size_t * axesB, const size_t naxes, const int normalize, float * fwork);
+             const size_t * axesB, const size_t naxes, const int normalize, int8_t  * fwork);
 void k2c_bias_add(k2c_tensor* A, const k2c_tensor* b);
 void k2c_flip(k2c_tensor *A, const size_t axis);
-float* k2c_read_array(const char* filename, const size_t array_size);
+int8_t * k2c_read_array(const char* filename, const size_t array_size);
 
 // Merge layers
 void k2c_add(k2c_tensor* output, const size_t num_tensors,...);
@@ -114,35 +115,39 @@ void k2c_maxpool1d(k2c_tensor* output, const k2c_tensor* input, const size_t poo
                    const size_t stride);
 void k2c_maxpool2d(k2c_tensor* output, const k2c_tensor* input, const size_t * pool_size,
                    const size_t * stride);
+//void k2c_maxpool2d(k2c_tensor* output, const k2c_tensor* input, const size_t * pool_size,
+//                   const size_t * stride);
 void k2c_avgpool1d(k2c_tensor* output, const k2c_tensor* input, const size_t pool_size,
                    const size_t stride);
 void k2c_avgpool2d(k2c_tensor* output, const k2c_tensor* input, const size_t * pool_size,
                    const size_t * stride);
+//void k2c_avgpool2d(k2c_tensor* output, const k2c_tensor* input, const size_t * pool_size,
+//                   const size_t * stride);
 
 // Recurrent layers
-void k2c_lstmcell(float * state, const float * input, const k2c_tensor* kernel,
-                  const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, float * fwork,
+void k2c_lstmcell(int8_t * state, const int8_t * input, const k2c_tensor* kernel,
+                  const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, int8_t * fwork,
                   k2c_activationType *recurrent_activation,
                   k2c_activationType *output_activation);
-void k2c_lstm(k2c_tensor* output, const k2c_tensor* input, float * state,
+void k2c_lstm(k2c_tensor* output, const k2c_tensor* input, int8_t * state,
               const k2c_tensor* kernel, const k2c_tensor* recurrent_kernel,
-              const k2c_tensor* bias, float * fwork, const int go_backwards,
+              const k2c_tensor* bias, int8_t * fwork, const int go_backwards,
               const int return_sequences, k2c_activationType *recurrent_activation,
               k2c_activationType *output_activation);
-void k2c_simpleRNNcell(float * state, const float * input, const k2c_tensor* kernel,
+void k2c_simpleRNNcell(int8_t * state, const int8_t * input, const k2c_tensor* kernel,
                        const k2c_tensor* recurrent_kernel, const k2c_tensor* bias,
-                       float * fwork, k2c_activationType *output_activation);
-void k2c_simpleRNN(k2c_tensor* output, const k2c_tensor* input, float * state,
+                       int8_t * fwork, k2c_activationType *output_activation);
+void k2c_simpleRNN(k2c_tensor* output, const k2c_tensor* input, int8_t * state,
                    const k2c_tensor* kernel, const k2c_tensor* recurrent_kernel,
-                   const k2c_tensor* bias, float * fwork, const int go_backwards,
+                   const k2c_tensor* bias, int8_t * fwork, const int go_backwards,
                    const int return_sequences, k2c_activationType *output_activation);
-void k2c_grucell(float * state, const float * input, const k2c_tensor* kernel,
-                 const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, float * fwork,
+void k2c_grucell(int8_t * state, const int8_t * input, const k2c_tensor* kernel,
+                 const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, int8_t * fwork,
                  const int reset_after, k2c_activationType *recurrent_activation,
                  k2c_activationType *output_activation);
-void k2c_gru(k2c_tensor* output, const k2c_tensor* input, float * state,
+void k2c_gru(k2c_tensor* output, const k2c_tensor* input, int8_t * state,
              const k2c_tensor* kernel, const k2c_tensor* recurrent_kernel,
-             const k2c_tensor* bias, float * fwork, const int reset_after,
+             const k2c_tensor* bias, int8_t * fwork, const int reset_after,
              const int go_backwards, const int return_sequences,
              k2c_activationType *recurrent_activation,
              k2c_activationType *output_activation);
