+++ apps/keras2c/include/k2c_merge_layers.c
@@ -6,7 +6,7 @@ Licensed under MIT License
 https://github.com/f0uriest/keras2c
  */
 
-#include "../string.h"
+#include <string.h>
 #include <stdarg.h>
 #include "k2c_include.h"
 
@@ -90,7 +90,7 @@ void k2c_average(k2c_tensor* output, const size_t num_tensors,...) {
 
     va_list args;
     const k2c_tensor *arrptr;
-    const float num_tensors_inv = 1.0f/num_tensors;
+    const int8_t  num_tensors_inv = 1.0f/num_tensors;
 
     va_start (args, num_tensors);
     memset(output->array, 0, output->numel*sizeof(output->array[0]));
